+++ torch_nerf/src/scene/__init__.py
@@ -1,2 +1,2 @@
 from torch_nerf.src.scene.primitives.primitive_base import *
-from torch_nerf.src.scene.primitives.primitive_cube import PrimitiveCube
+from torch_nerf.src.scene.primitives.cube import PrimitiveCube
