+++ torch_nerf/src/scene/primitives/cube.py
